converts int to double

+++ Envelopes/Data/NewEnvelope.swift
@@ -11,10 +11,10 @@ import Foundation
 struct NewEnvelope {
 
     var name: String?
-    var recurringAmount: Int = 0
-    var startingAmount: Int = 0
+    var recurringAmount: Double = 0
+    var startingAmount: Double = 0
     var periodicity: Periodicity = .monthly
-    var goal: Int = 0
+    var goal: Double = 0
 
     var isReady: Bool {
         return name != nil &&
